+++ parachain/Cargo.toml
@@ -16,7 +16,7 @@ members = [
     "pallets/eth-app",
     "pallets/erc20-app",
     "pallets/commitments",
-    "pallets/token-dealer",
+    "pallets/transfer",
     "node",
     "runtime",
 ]
