+++ parachain/runtime/Cargo.toml
@@ -46,7 +46,7 @@ polkadot-parachain = { git = "https://github.com/paritytech/polkadot.git", branc
 xcm-handler = { git = "https://github.com/paritytech/cumulus.git", branch = "rococo-v1", default-features = false }
 
 artemis-core = { path = "../primitives/core", default-features = false }
-artemis-token-dealer = { path = "../pallets/token-dealer", default-features = false }
+artemis-transfer = { path = "../pallets/transfer", default-features = false }
 artemis-xcm-support = { path = "../primitives/xcm-support", default-features = false }
 bridge = { path = "../pallets/bridge", package = "pallet-bridge", default-features = false }
 dispatch = { path = "../pallets/dispatch", package = "artemis-dispatch", default-features = false }
@@ -99,7 +99,7 @@ std = [
     "eth-app/std",
     "erc20-app/std",
     "commitments/std",
-    "artemis-token-dealer/std",
+    "artemis-transfer/std",
     "artemis-xcm-support/std"
 ]
 test-e2e = []
